`PsqlDos`: Make hash reset migrations more explicit
The `django_0039_reset_hash.py` and `e797afa09270_reset_hash.py`
migrations reset the hash for all nodes by deleting the extra in which
it is stored. The migrations use the `drop_hashes` function from the
`aiida/storage/psql_dos/migrations/utils/integrity.py` module. At the
time, and currently, the key used to store the hash in the extras is
`_aiida_hash`. This key is hard-coded in the `integrity.py` module but
it should really be part of the migrations themselves if the function is
to be reused. If in the future the key is changed, the migration should
specify which key should be used.