`PsqlDos`: Add migration to remove hashes for all `CalcJobNodes`
The algorithm for computing the hashes of `CalcJobNodes` contained a bug
which was fixed in 685e0f87d7c571df24aea8f0ce21c6c45dfbd8a0. This means
however that all existing hashes for `CalcJobNodes` are inconsistent
with the new algorithm. This would mean that valid cache hits would be
missed, and worse, different calculatons could end up with the same hash
and mistakingly be cached from one another.

The migration drops the `_aiida_hash` extra which is used to store the
node's hash. A warning is emitted to notify the user, suggesting to run
`verdi node rehash` to recompute the hash of all `CalcJobNodes`.