`SinglefileData`: Fix bug when `filename` is `pathlib.Path`
It would raise an exception when the node was stored since it would
compare the `filename` against the list of objects, but the latter would
be instances of a string instead of `pathlib.Path` and the validation
would fail. On top of that, the difference in type was not visible as
the cause in the error message, making for a cryptic error.