`DaemonClient`: Clean stale PID file in `stop_daemon` (#6007)
The daemon client was recently refactored to only clean the stale PID
file in the `start_daemon` command. Before it was done when asking the
status which was considered an unexpected side-effect.

Sometimes, the user wants to make sure the daemon is no longer running.
Currently, the status would raise a warning if a stale PID file is found
suggesting the user to start the daemon to fix it. However, this is
counterintuitve and not desirable if the goal is for the daemon to be
stopped.

The `stop_daemon` method is updated to also clean any potentially stale
PID files. The error message is updated to suggest to either start or
stop the daemon to return it to a nominal state.

Co-authored-by: Sebastiaan Huber <mail@sphuber.net>