ORM: Fix problem with detached `DbAuthInfo` instances
The idea is that somehow, the `AuthInfo` instance that is kept as a
reference by the `JobsList` goes stale at some point when a `CalcJob`
requests a transport through the `JobManager`. The fix is to use the
explicit `AuthInfo` that is passed when `request_job_info_update` is
called and not the one that was used to construct the original
`JobsList` instance.