Correct instructions to rehash nodes after migration (#6028)
After a migration that can affect the hash of the migrated node, the
instructions to rehash the node are provided. However, these instructions are
currently incorrect and likely to lead to confusion for the user.

Here the instructions are corrected both when a specific entry point requires
rehashing or not.