`InteractiveOption`: Fix validation being skipped if `!` provided
The `InteractiveOption` reserves the exclamation point `!` as a special
character in order to "skip" the option and not define it. This is
necessary for options that are _not_ required but that do specify a
defualt. Without this character, it is impossible to _not_ set the
default as, if the user doesn't specify anything specific and simply
presses enter, the default is taken, even if the user did not want to
specify any specific value.

The problem with the implementation, however, is that when `!` was
provided, the option would return `None` as the value, bypassing any
validation that might be defined. This would make it possible to bypass
the validation of a required option.

The solution is to, when `!` is provided for an interactive option, it
is translated to `None` and is then processed as normal, validating it
as any other value.