Docs: Fix reference target warnings related to `flask_restful`
The REST API dependency `flask_restful` released `0.3.10` recently which
caused the RTD build to start warning about the `Api` and `Resource`
classes no longer being findable. The classes are still importable and
looking at the changes with `0.3.9` it is not clear that the imports
should have changed, so for now these are added to the nitpick ignore
list.