`DirectScheduler`: Add `?` as `JobState.UNDETERMINED`
When running a process via `engine.run` on a MacOS `localhost` using the
`core.direct` scheduler, warnings are often raised regarding an
"unrecognised job state `?`'. This is related to the fact that on MacOS the `ps`
command is sometimes not able to determine the process state temporarily, and
prints a `?` in the `STAT` column, see:

https://apple.stackexchange.com/q/460394/497071

Here we simply add the `?` as one of the possible outputs of the `ps` commmand
in the `_MAP_STATUS_PS` dictionary, mapping it to `JobState.UNDETERMINED`.