Process control: Warn instead of except when daemon is not running (#6047)
The `play_processes`, `pause_processes` and `kill_processes` methods of
the `engine.processes.control` module would raise if the daemon is not
currently running. This instinctively make sense, because in order to be
able to interact with a process through an RPC call, which is what these
methods do, it has to be "live" with some runner. This normally means a
daemon worker, so if the daemon is not running, it is to be expected
that the process cannot be reached. However, runners can also be started
manually, for example through `verdi daemon worker` or within an
interactive shell when a process is "run" instead of submitted.

The `verdi process pause/play/kill` commands actually covered for this
case and would also check for the daemon to be running, but instead
would log a warning if not the case. Since they call through to the
control module, an exception would be logged nevertheless.

The control module now also logs a warning, instead of raise an
exception. This allows the CLI command to remove the explicit check for
the daemon running with the `only_if_daemon_running` decorator as it can
rely on the control module performing the check.