Process control: Change language when not waiting for response (#6048)
When playing, pausing or killing a process with `wait=False`, the
response would be like: `played/paused/killed Process<ID>`. This
suggests to the user that the action was completed, but actually,
without waiting, the RPC was simply fire-and-forget and so it may take a
while for the action to actually be completed, if at all.

Therefore the language could be misleading, especially if the daemon is
under heavy load and cannot execute the RPC relatively quickly. The
language is changed to indicate that a request was sent, which should
convey that it is not necessarily completed already.