Fix log messages being logged twice to the daemon log file
In 276a93f3a60546674d0b18220f6415ed33716c89 the logging configuration
logic was modified to respect the log level being changed through the
`--verbosity` option when invoking `verdi`. The changes added the `cli`
handler to all loggers.

This code was also being executed when a daemon worker started up and
called `aiida.common.log.configure_logging` causing all logging to end
up twice in the daemon log file.

The assertion of the `test_duplicate_subscriber_identifier` test is
updated since the expected error message contained the `Error:` literal
which is the log level at which the message is logged and is therefore
formatted based on the loggers formatter, which is configurable. In this
case, it is actually formatted as `[ERROR]` which was causing the test
to fail.