`QueryBuilder`: Deprecate `debug` argument and use logger
The `QueryBuilder` defined the `debug` flag, which when set to `True`
would cause certain methods to print additional debug information. This
should go through the logging functionality, however, so the print
statements are replaced by a `logger.debug` call.

Since the debug messages can now be activated by configuring the logger
log level, the `debug` argument is no longer needed and is deprecated.