CLI: Added `compress` option to `verdi storage maintain` (#5540)
By default, the disk-object store, which is used by the default storage
backend `PsqlDosStorage`, does _not_ compress files when joining them
into pack files during the maintenace operation. This behavior can be
toggled through a flag in the interface, but this is not exposed in the
`verdi storage maintain` and so AiiDA users currently cannot benefit
from file compression.

Moreover, in the current implementation of the disk-object store, once
packed with or without compression, it cannot be repacked with a
different setting.

Here the `--compress` option is added to `verdi storage maintain` which
can change the default behavior of not compressing when packing loose
files.

Note that this is a storage backend dependent option that not all storage
backends may support, but currently the spread of alternative storage
backends in the AiiDA ecosystem is very limited and it is unlikely that
a fully dynamic interface for `verdi storage maintain` is necessary in
the near future. If this need arises in the future, the `--compress`
option might have to be deprecated, but this is an acceptable cost to
prevent from having to implement a fully dynamic CLI interface at this
point in time.