`verdi process status`: Add `call_link_label` to stack entries
The `metadata.call_link_label` input can be set to define a custom link
label for call links when launching a subprocess, instead of the `CALL`
default. The `aiida.cmdline.utils.ascii_vis.calc_info` function, which
is called by `verdi process status` to produce a visual representation
of the call stack of a process, is updated to include this call link
label.