Config: Add option to change recursion limit in daemon workers (#6044)
Add the `daemon.recursion_limit` config option which is used by a daemon
worker to set the recursion limit using `sys.setrecursionlimit`. This is
to help with the current problem where heavy workloads with many process
functions can create deep stacks which would result in `RecursionError`
exceptions being thrown. The default is set to 3000, increased from the
built-in default of 1000. This value has precedent as it is also used by
`jedi` which is used for `ipython`.