Devops: Update code to fix mypy error codes
The `requirements/requirements-3.9.txt` was recently updated which is
also used for the `pre-commit` CI job and so new `mypy` errors cropped
up. The biggest change is changing the `verdi` root command from
`click.command` to `click.group`. `mypy` was rightly claiming that the
decorated `verdi` function did not have attributes used to create
sub-groups and sub-commands as a `Command` doesn't have those.