`CalcJob`: Ignore file in `remote_copy_list` not existing
If a statement in the `remote_copy_list` contained an explicit filename
to copy, i.e., without wildcards, the transport would raise an `OSError`
if the file didn't exist. This exception would be reraised and the
exponential backoff mechanism would be triggered. However, in the case
that the file does not exist, it really is not a transient problem and
the operation would be guaranteed to fail in all retries as well.

In the previous commit, the transports were changed to raise the more
specific `FileNotFoundError` if the source in a `copy` operation does
not exist. This exception is now caught separately in the
`upload_calculation` method and instead of reraising, a warning is
logged and the exception is swallowed.

This behavior falls more in line with other file handling operations of
the `CalcJob`, for example in the retrieval of files. There, also, a
missing file does not trigger an exception but is simply logged with a
warning.