Fix the tests
The tests were failing because `mayavi` is not installed and so was
raising an `ImportError`. In addition, the tests would hang because the
show command would actually open a window to display the generated image
for certain formats. Since the purpose of these CLI tests are not
actually to test the implementation of the visualization but just the
CLI command, the show function is mocked with a function that simply
checks that the expected arguments are passed, based on the options that
are passed to the command line invocation.