Adapt functionality
Before, if `--create-db` was set on and the user provided a
name for the database that already existed, AiiDA would try
then with appending `_1`, `_2`, ..., `_N` at the end of the
name. This was the default behavior when creating the new
database was the only option.

Now the command will except if the name is provided but the
database already exist, and it is up to the user to provide
an alternative name. The only exception is if no name was
provided and AiiDA is trying to use the default, in which
case it will again try to append `_N` if the name already
exists.