`ArrayData`: Allow defining array(s) on construction
Currently, the constructor does not allow to define any arrays to set
when constructing a new node, so one is forced to multi line code:

    node = ArrayData()
    node.set_array('a', np.array([1, 2]))
    node.set_array('b', np.array([3, 4]))

This commit allows initialization upon construction simplifying the code
above to:

    node = ArrayData({'a': np.array([1, 2]), 'b': np.array([3, 4])})

Note that it is also possible to pass a single array to the constructor,
in which case the array name is taken from the `default_array_name`
class attribute.

For backwards compatibility, it remains possible to construct an
`ArrayData` without any arrays.