`RemoteData`: Add the `is_cleaned` property (#6101)
This is a convenience method that will return the `KEY_EXTRA_CLEANED`
extra, which is set to `True` when the `clean` method is called. The
`is_empty` method is also updated to use this new property and shortcut
if set to `True`. This saves the method from having to open a transport
connection.