Remove `with_dbenv` use in `aiida.orm`
This forces the import of `aiida.cmdline` in `aiida.orm` which doesn't
just slow down, but also is conceptually wrong. The problem of the
`with_dbenv` decorator is also that it cannot be imported inside a
method to avoid the import cost when importing `aiida.orm` but has to be
imported at the top in order to be used.