ORM: Add the `Entity.get_collection` classmethod
This is an alternative for the `collection` class property. Where the
`collection` property uses the current default storage backend, the
`get_collection` allows to specify another specific backend.

The original design intended to support this use-case by allowing a
`Collection` instance to be "called" with a specific backend:

    Entity.collection(backend)

The `.collection` returns a `Collection` instance which is then called
with passing the `backend`, which would return a new `Collection` for
the same entity type, but with the other backend. However, this doesn't
always work, because the `collection` property will load the backend
from the default profile, which will except if not loaded. Although this
scenario is unlikely for normal usage, application developers may use
AiiDA's API with multiple active backends where no default profile is
defined.

The reason for a new method instead of changing the `collection`
property is that that would be backwards incompatible.