Dependencies: Add compatibility for `pymatgen>=v2023.9.2`
As of v2023.9.2, the ``properties`` argument of the `Specie` class is
removed and the ``spin`` argument should be used instead. See:
https://github.com/materialsproject/pymatgen/commit/118c245d6082fe0b13e19d348fc1db9c0d512019

The ``spin`` argument was introduced in v2023.6.28. See:
https://github.com/materialsproject/pymatgen/commit/9f2b3939af45d5129e0778d371d814811924aeb6

Instead of removing support for versions older than v2023.6.28 the code
is updated to be able to deal with the new version where `properties` is
no longer supported.