Use fastjsonschema to validate user configuration
We use jsonschema to validate user configuration
upon each verdi invocation. The problem is that importing
`jsonschema` package is very slow.

In this PR we use the fastjsonschema instead. This
package uses code generation for generating fast python
code based on a given schema.
To fully take advantage of this and to avoid the compilation step,
the code is pre-compiled and checked-in. The disadvantage is that
when the schema changes, the code needs to be re-generated as well.