Tests: Fix flaky work chain tests using `recwarn` fixture (#6112)
The tests were often failing because the `recwarn` fixture contained
two records instead of one. The reason is that elsewhere in the code a
`ResourceWarning` is emitted because an event-loop is not closed when
another one is created. Until this is fixed, the assertion is updated to
not check for the number of warnings emitted, but specifically to check
the expected warning message is present.