Tests: Make `PsqlDosStorage` profile unload test more robust (#6115)
The `test_unload_profile` test verifies that if a loaded profile is
unloaded, it properly relinquishes of the session that is maintained by
sqlalchemy. It did so by checking that after unloading, there were no
sessions being referenced. However, this would fail sometimes, because
another session may still be held on to, even though that session had
nothing to do with the test.

A more robust test is simply to check that after unloading, there is
exactly one less session being held on to.

Cherry-pick: 1c72eac1f91e02bc464c66328ea74911762b94fd