Refactor: Delay import of heavy packages to speed up import time (#6116)
The importing of packages `urllib`, `yaml` and `pgsu` are moved from
top-level to inside the scopes where they are needed. This significantly
improves the load time of the `aiida` package and its subpackages.

The import of `Transport` in the `aiida.engine` package also slows down
imports, but it is only used for type checking, so its import is placed
inside the `if TYPE_CHECKING` guard.

Finally, the `DEFAULT_DBINFO`, `Postgres` and `PostgresConnectionMode`
objects of the `aiida.manage.external.postgres` package are no longer
exposed on the top-level as this also slows down imports. This is a
breaking change technically, but these resources should not be used by
downstream packages.